Consolidate to using just the ListActivity (i.e., remove the Service), since the service isn't needed.  Add progress messages.  Loading arrival times is  really slow, need to examine why.
