+++ onebusaway-android/src/main/res/values/strings.xml
@@ -319,4 +319,6 @@
     <string name="find_my_bus_voice_trigger">find my bus</string>
     <string name="stop">Stop</string>
     <string name="read_aloud">Read aloud</string>
+    <string name="finding_closest_stop">Finding closest stop…</string>
+    <string name="getting_arrival_times">Getting arrival times…</string>
 </resources>
